refactor(native): close review-pass data-loss, shutdown, and cleanup gaps
Follow-up to the native hardening review. Fixes real data-loss and
shutdown bugs plus a batch of lower-risk cleanups.

- recognizer: write record.json last (after prediction.json/trainee.jpg)
  and skip trainee.jpg when the trainee-icon crop is empty. Previously a
  missing trainee reference point produced a degenerate ROI whose OOB
  throw dropped the whole (otherwise valid) record, and record.json was
  written first so a mid-write throw left it without its sidecars -- which
  the loader treats as a decode failure and deletes.
- json_util: add EXTENDED_JSON_TYPE_ENUM_STRICT, which throws on an
  unknown enum name instead of the stock NLOHMANN_JSON_SERIALIZE_ENUM
  behavior of silently mapping it to the first enumerator. Apply it to
  RecordType and LayoutAnchor so a config/record skew fails loudly rather
  than silently remapping to Standard / ScreenStart.
- cli: replace exit(1) with a return-code path so a failing subcommand
  unwinds its stack (destroying/joining its local runners) and still runs
  spdlog::drop_all(). Make stitch/recognize/video self-terminate by
  joining the event loop once the pipeline goes quiet, instead of spinning
  in isRunning() forever (capture stays open-ended).
- native_api: joinEventLoop() now returns without logging when nothing is
  running. It also runs from ~NativeApi at process exit (atexit), after
  spdlog::drop_all() has destroyed the default logger, so logging there
  dereferenced freed logger state and crashed -- newly reachable now that
  the one-shot subcommands exit gracefully instead of being killed.
- frame_distributor: drop the permanently-null on_no_target sender/branch.
- frame: move the header's anonymous namespace (BGR/linspace/asBGRRange)
  into a named frame_impl namespace with inline functions, avoiding a
  per-TU internal-linkage copy and a distinct BGR type per TU.
- fix signed/unsigned index loops in the recognizer and scraper.

Verified: Debug build + all 15 unit tests pass; `build` round-trips the
strict LayoutAnchor enum; `recognize` now self-terminates cleanly
(exit 0) where it previously hung and, once exiting, crashed at atexit.

+++ native/src/chara_detail/chara_detail_scene_scraper.cpp
@@ -11,7 +11,7 @@ bool closeEnough(const std::vector<double> &a, const std::vector<double> &b, dou
         return false;
     }
 
-    for (int i = 0; i < a.size(); i++) {
+    for (size_t i = 0; i < a.size(); i++) {
         if (std::abs(a[i] - b[i]) > threshold) {
             return false;
         }
